requirements.txt
boto3==1.34.127
botocore==1.34.127
gTTS==2.5.3
moviepy==1.0.3
playwright==1.44.0
praw==7.7.1
prawcore~=2.3.0
requests==2.32.3
rich==13.9.2
toml==0.10.2
translators==5.9.3
pyttsx3==2.98
Pillow==10.4.0
tomlkit==0.13.2
Flask==3.0.3
clean-text==0.6.0
unidecode==1.3.8
spacy==3.7.6
torch==2.4.1
transformers==4.44.2
ffmpeg-python==0.2.0
elevenlabs==1.8.1
yt-dlp==2024.10.7
numpy==1.26.4
azure-cognitiveservices-speech==1.41.1
vosk==0.3.45